requirements.txt
streamlit>=1.24.0
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0